Extract CQL execution logic to CqlExecutor
