+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CqlExecutor.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2016 Palantir Technologies
+ *
+ * Licensed under the BSD-3 License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://opensource.org/licenses/BSD-3-Clause
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.palantir.atlasdb.keyvalue.cassandra;
+
+import java.net.InetSocketAddress;
+import java.nio.ByteBuffer;
+
+import org.apache.cassandra.thrift.Compression;
+import org.apache.cassandra.thrift.ConsistencyLevel;
+import org.apache.cassandra.thrift.CqlResult;
+import org.apache.thrift.TException;
+
+import com.google.common.base.Charsets;
+import com.palantir.common.base.Throwables;
+
+public class CqlExecutor {
+    private CassandraClientPool clientPool;
+    private InetSocketAddress host;
+    private ConsistencyLevel consistency;
+
+    public CqlExecutor(CassandraClientPool clientPool, InetSocketAddress host, ConsistencyLevel consistency) {
+        this.clientPool = clientPool;
+        this.host = host;
+        this.consistency = consistency;
+    }
+
+    public CqlResult execute(String query) {
+        ByteBuffer queryBytes = ByteBuffer.wrap(query.getBytes(Charsets.UTF_8));
+        return executeQuery(queryBytes);
+    }
+
+    private CqlResult executeQuery(ByteBuffer queryBytes) {
+        try {
+            return clientPool.runWithRetryOnHost(host, client -> client.execute_cql3_query(queryBytes, Compression.NONE, consistency));
+        } catch (TException e) {
+            throw Throwables.throwUncheckedException(e);
+        }
+    }
+}
